+++ src/app/sat/anytime_sat_clause_communicator.hpp
@@ -46,7 +46,7 @@ private:
     void learnClauses(const std::vector<int>& clauses);
     void sendClausesToChildren(const std::vector<int>& clauses);
 
-    std::vector<int> merge(const std::vector<std::vector<int>*>& buffers, int maxSize);
+    std::vector<int> merge(const std::vector<std::vector<int>*>& buffers, size_t maxSize);
     bool testConsistency(const std::vector<int>& buffer);
 };
